+++ treedoc/tests/__init__.py
@@ -1,7 +1,5 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """
-Created on Sun Apr 21 21:14:43 2019
-
-@author: tommy
+Package for testing the treedoc package.
 """
